feat: Add createTransaction and getTransactionsByBankId functions
This commit adds the createTransaction and getTransactionsByBankId functions to the transaction.actions.ts file. The createTransaction function allows the creation of a new transaction document in the database, while the getTransactionsByBankId function retrieves transactions based on the bank ID. These changes enhance the functionality of the application by providing the ability to create and retrieve transactions.